[WIP] Use latest version of the CDI
This patch updates the tests so that it uses kubevirt-ansible to install
the latest supported version of the CDI.
The latest supported version through kubevirt-ansible is v1.1.1.
The latest release is v1.2.0.

It also removes the CDI manifest that was being used to install v0.5.0.

Requires: https://github.com/kubevirt/kubevirt-ansible/pull/399