Fix multus config
The config in kubevirt-ansible sets the multus CNI as flannel, but the 
CNI that kubevirt-ansible installs is weave-net.

This change updates the config to use weave-net. It should be removed 
when kubevirt-ansible has been updated to deploy flannel instead of 
weave-net or the correct multus cni configuration if weave-net is indeed 
the desired default.