Fix symbolic links in kubevirt-ansible
The symbolic link at kubevirt-ansible/playbooks/role wasn't being 
preserved when the directory was copied to the EBS volume. Instead of a 
link, a copy of the directory represented by it was reproduced, 
duplicating code, and making it confusing as to where to edit roles.

As a solution, we now tar the contents of the kubevirt-ansible, copy the 
tar file over to the EBS volume, and extract it.