Update tests to use commands from kubevirt.io website
We want tighter integration between the commands the users are using in 
the labs and the commands we execute in the CI tests.

Parts of the commands used in the kubernetes lab on kubevirt.io have 
been templated out into the labs_kuberentes_variables.yml file. This 
change updates the tests to read in the variables from this file and 
then use the variables to form test commands.