Revamped way of creating images on Cloud vendors done by Joel Davis from https://github.com/joeldavis84/kvio-lab-images
Changed way of work by a more standard one, using a Jenkins pipeline to Build/Upload/Distribute between GCE and AWS, also spreading the AWS AMI through some regions. This is not a finished work, for that we have a TO-DO list on README file in order to improve the process.